Updated openstack/openstack Project: openstack/neutron  728a0fd356bd8a98d985aca22f8faeabf9fc36b7
Retry port update on IpAddressAllocationNotFound

If a port update and a subnet delete interleave, there is a
chance that the IPAM update operation raises this exception.
Rather than throwing that up to the user under some sort of
conflict, bubble up a retry instead; that should bring things
back to sanity.

Closes-bug: #1622616

Change-Id: Ia8cac09349d4cb722737bdf0bec6c54b9e77f31d