Updated openstack/openstack Project: openstack/nova  59a0c0775a3e19ac89fa5d3c8d11766e7eb48316
docs: Clarify sections & note on filter scheduler

Right on now the Filter Scheduler page it is difficult to find where the
section is on configuring and writing your own. The page is very long.
So this adds some section headers. In addition a note is added to
clarify that when you write your own filter "all_filters" doesn't
include "write your own", both available and default settings need to be
changed.

Change-Id: I99878d3b4dc839ff87bc581e61c85e4ae2c66b2c