Updated openstack/openstack Project: openstack/nova  4d789f94e0a0daf44d35e94cf766353d768f2041
[placement] 404 responses do not cause exception logs

Prior to this change, a 404 caused by a not possible route being
requested (such as /placement/hello) would result in a proper 404
response but also a log message about an uncaught exception. This
made it appear that a 404 response is unexpected and something that
needs exceptional attention. It doesn't. This change makes sure
that 404s propagate up the stack in a clean way and are logged only
by the existing INFO log used for all requests.

Change-Id: I21f9411bd5154eddc582d5f38891be747de874b3
Closes-Bug: #1626489