Updated openstack/openstack Project: openstack/nova-specs  51b0bb1b674e1ee01d825fcb3b12c133fd7ea2f7
Add Keystone project validation to quota and flavor management

Currently, Nova does not check the validity of what the user provides for the
project id value in quota and flavor management activities. This leads to
inaccurate quota entries in the database, causing the user to think Nova isn't
reporting quota information accurately. The solution proposed in this blueprint
is to add project id validation specifically to quota and flavor management
through Keystone, to provide the user with an error if their specified project
id is invalid, and to ensure quota and flavor management data is not incorrectly
modified.

bp validate-project-with-keystone
Previously-approved: newton

Change-Id: Ib5d22b2c52d04114d91bf5313e72f7e9a13abe92