Updated openstack/openstack Project: openstack-infra/project-config  f2d1cf71ed213efd3057e6b7a51a82167d447e77
Merge zuul-git-prep-upper-constraints/zuul-release-git-prep-upper-constraints

Both builders do basically the same thing: Check out a repository and
the requirements repo and make the upper-constraints file available.

Problem is that zuul-git-prep-upper-constraints cannot use in
tag/release queues. The builder
zuul-release-git-prep-upper-constraints was written for this. It can be
used in all queues, so let's use that one instead of having two
builders.

The change removes zuul-git-prep-upper-constraints, renames
zuul-release-git-prep-upper-constraints to
zuul-git-prep-upper-constraints and changes usage of
zuul-release-git-prep-upper-constraints to
zuul-git-prep-upper-constraints.

So, we now have one builder under the
name zuul-git-prep-upper-constraints that can be used everywhere.

Change-Id: Ifbebd0f8a7b265dfe517b4c180502580b4b0b55f