Updated openstack/openstack Project: openstack/nova  08338dba5e2d9a127dac8a52b12cf124821b0a63
Clean up stdout/stderr leakage in cmd testing

This cleans up stdout/stderr leaking in cmd testing by using the
pattern of MonkeyPatching stdout/stderr to a StringIO object where
appropriate.

The OutputStreamCapture fixture was using in one place, but that
actually is just about moving the output to the subunit stream. That
output will eventually be dumped at the end of the runs. When we
really don't want that to happen, we need to capture it directly.

Change-Id: Ib788ceccd72677d4602b6e6b77e74a4abff8fec5