Updated openstack/openstack Project: openstack-infra/project-config  9e3bff914df259ac10f3b0ef1077d523bd7383b3
Add missing root-marker for afs-publishing

Add missing root-marker to jobs that publish to docs.openstack.org.

This is so that we can create a list of directories to ignore when
we rsync documentation builds to a publishing site.  Some builds
are published as subdirectories underneath the results of superior
builds.  This allows those superior builds to be rsynced without
removing the other builds that were published under them.

Change-Id: I3c5555829bd064675b2fded82cdafaa0916294dd