Updated openstack/openstack Project: openstack/neutron  d05cf4d7fdf18adf981dbc7938f549c69eed513d
Send a callback before attaching a subnet to a router

Attaching a subnet to a router may be forbidden in some cases.
To prevent this attachment, we need to send a BEFORE_CREATE event
with "router_interface" as resource, and reraise any exception
that tells neutron a third party code forbid this attachment.

Change-Id: Ia71e0c5e9e3e073053ead7e1e7c3040087f5ee13
Closes-bug: 1537091