Updated openstack/openstack Project: openstack/neutron  83284c4e24db96b184b5ef3ce74b3fc1f39a6f39
Check for l3 agents count before router update

When updating a legacy router to be an HA, it's possible that there are
less than the minimum required l3 agents available. In this case an
exception will be raised to indicate this, but some of the changes will
already be in committed. The effect is that the router will be in an
invalid partial state between legacy and HA.

This patch simply checks that there are enough l3 agents before starting
to convert the router to HA. This adds a single DB query for each
router-update from legacy to HA, but saves a bunch of others in case the
transition should not take place.

Closes-Bug: #1609665
Change-Id: I566b17003f6d9b4c8205c0586f742795012da29c