Updated openstack/openstack Project: openstack/python-openstackclient  f19240fc29c542726acd11a7f71f307be219695f
Add default limit for container/object

Default container name length less than or equal to 256 in link[1],
as the same time,default object name length less than or equal to 1024
in link[2]. Thereforce, I check the length of container and object in take_action.
and if it's greater than 256/1024 I warn the user.

[1] https://github.com/openstack/swift/blob/master/swift/common/constraints.py#L39
[2] https://github.com/openstack/swift/blob/master/swift/common/constraints.py#L35

Change-Id: I304b77cbc464eaba041321654cc29248cbe4b9a6