Updated openstack/openstack Project: openstack-dev/devstack  71afa252500b73a03abc046fbcc0c13d9847cfc5
Don't make root CA if it exists

To support multinode testing where we just copy the CA to all the
instances don't remake the CA if it already exists.

The end result is that you can trusty a single chain and all your
clients will be happy regardless of which host they are talking to.

Change-Id: I90892e6828a59fa37af717361a2f1eed15a87ae4