Updated openstack/openstack Project: openstack/neutron  c64b6295d9120372ab0c600c3cb087b6f2cbfc3c
Process OVS trunk bridges associated to VM deletes

If a VM is deleted while the OVS agent is down, upon
restart the OVS trunk bridge is deleted successfully
but the patch ports integration bridge side are left
behind.

This patch makes sure the bridge as well as its peers
are cleaned up.

Closes-bug: #1623708

Change-Id: I3f010755fdb6501753d20357ad2cd0d6c02cf22a