Updated openstack/openstack Project: openstack-infra/nodepool  2d1068426d8bc5c74831ddbd2a8e565f723bc912
Add image id to DIB image upload name params

It would be handy to be able to notice that a disk image is already
uploaded into a cloud and just treat that as completed. Shade does most
of the work for us here we just have to make sure we ask it to create
the image using the same name again. To do this add image_id to the
image that can be interpolated into the image name. This id is fixed for
each build of the image unlike the current timestamp which is based on
time now.

Change-Id: Icedbba40adcf9b3e45bdbc55fca7729c68cfd0b9