Updated openstack/openstack Project: openstack-infra/project-config  fb035b15e9e5b61baea559af4a82d30ac13c6e57
Updated OSIC project for HW specific flavors

In the OSIC we'd like to split workloads such that they're running on
hardware specific flavors. This is being done so that we can gather
metrics on our various hardware profiles. The initial commit has a
low 24 node quota however once we establish that everything works as
expected we intend to spread the quota out evenly and remove the
original cloud1-config entry.

Change-Id: I4f12c9013040a49c17e5a636dd3e0d719a2ace3e
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>