Updated openstack/openstack Project: openstack/governance  5d68af740ad28eee60af8597fbe7421bda9a0e08
Remove Astara from governance

The Astara team believes the community's best interest are served by
removing the Astara from TC governance. Most of the projects's core team has
moved to different roles and there is not a ready set of maintainers available
to replace them and continue the project in its current form.  Retiring the
project will enable interested members of the community to reorganize
and restructure and if desired reapply for goverance in the future.

The Newton release will be the last release by the current team and will
provide time for any deployments to migrate to another solution.

Change-Id: Ic8e0e3115a379bdc876f6a064519b233515bcd3a