Updated openstack/openstack Project: openstack/governance  6cfd158f276973a255a8a849ebc57878160d03f0
Remove OpenStack-Salt from governance

OpenStack Salt project has missed several deadlines for various
reasons and was probably way too early for them to be in the
Big Tent. Several contributors to the project were in a startup
that recently got acquired as well. When they figure out their
next steps and understand what it means to be under the "Big Tent"
they are welcome to re-propose adding them back.

Change-Id: I6e55b7404a53c8326c3a3eed065bd6c91db0e1ec