Updated openstack/openstack Project: openstack/governance  3a1e2011b2761e3445a8925919f392d5a32ecd68
Tag neutron-lib as tc-approved-release

Neutron-lib was created in late 2015 by the Neutron team with the intent
of splitting out stable, reusable library code from the main Neutron
repository. [1]

As such, neutron-lib is an important piece of under-the-hood machinery
for how Neutron functions going forward.  Some or all of it may be a
reasonable candidate for inclusion in future DefCore Interoperability
Guidelines as a Designated Section.  Although as a library project it
does not expose any Capabilities to end users of OpenStack clouds
directly via an API, it does provide code that is or may in the future
be invoked by Capabilities that are required by Interoperability
Guidelines.

At present, the DefCore Committee would be unable to consider adding
part or all of neutron-lib as a Designated Section to future Guidelines
because it isn't designated as part of the TC-approved release. [2]
This means (as an extreme example) that a vendor could entirely replace
neutron-lib with it's own independently created code and still have the
resulting product bear the OpenStack trademark.  By tagging neutron-lib
as tc-approved-release, the DefCore Committee would be allowed to
consider adding Designated Sections to future Guidelines that include
neutron-lib, thereby making the community developed code required.

[1]
http://specs.openstack.org/openstack/neutron-specs/specs/liberty/neutron-lib.html
[2]
http://git.openstack.org/cgit/openstack/defcore/tree/doc/source/process/2016A.rst#n86

Change-Id: I6befe442c31b812f3d3b17b21ac844be96a419b2