Updated openstack/openstack Project: openstack/governance  25086b5ea9a109b1cf5d492202b3b8aba9182de9
Write down OpenStack principles

Beyond the Four Opens, the OpenStack community (and the Technical
Committee as its direct representation) operates on a lot of
unwritten principles, which were transmitted by oral tradition
by the elders and appear from time to time as a snippet of wisdom
in the middle of a TC meeting.

It is more than time for us to document those principles, to make
them clearer to everyone and harder to ignore. To give credit where
due, the bulk of this document was written by Monty Taylor, with
input from a number of other folks including (but not limited to)
the list of co-authors below.

Change-Id: I0134b0aa9b512ee61fbb908d23e7260ff2491bba
Co-Authored-By: Monty Taylor <mordred@inaugust.com>
Co-Authored-By: Doug Hellmann <doug@doughellmann.com>
Co-Authored-By: John Garbutt <john@johngarbutt.com>