Updated openstack/openstack Project: openstack/tempest  a6a58a743f8fc573afe97ad61c7401e7d8ee77e1
Fix subunit-trace output with tempest run

Now that we're using tempest run in tox it became clear that the
subunit-trace arguments we were using in tempest run did not match what
we were previously using in pretty_tox.sh. This makes debugging failures
slightly more difficult than before because the failures aren't printed
at the end of the run. This commit adds the necessary flags to tempest
run calling of subunit-trace to have the same behavior as before.

Change-Id: Ib066acca792d4f357763235ef4593062e9978f0b