Updated openstack/openstack Project: openstack/python-ironicclient  e512550ab7d61b8bdafd50022dfa8b80a8502922
Fix python3 compatibility when HTTP Error are returned

When trying to contact Ironic with a bad token, Forbidden exception
should be raised, in python3 a TypeError is raised due to json
lib being unable do decode a bytes object.

In order to be really python3 compatible, the json lib was replaced
with oslo.serialization module jsontuils since it's the recommended
migration to python3 guide. This is to ensure that data coming from
the requests lib can be read even if it's not string any more but
bytes.
https://wiki.openstack.org/wiki/Python3

Change-Id: I27540f58e31817d4de604334bc4c62899d82f4cc
Closes-Bug: #1629068