Updated openstack/openstack Project: openstack/requirements  291a5ff9fb09d14e5ea88e7786b10d9fb9d59b56
Update wsgi-intercept to 1.4.1

1.4.1 adds (optional) support for checking that response headers
sent by the intercepted application are in the proper form (`str`
objects in python 2 and python 3, thus bytes in one and not the
other). This is hard to test in the code that uses wsgi-intercept
because most HTTP libraries will normalize the response headers
to a convenient type.

This functionality will be used in the tests of the placement
API in nova.

Change-Id: I3b8aabda929fe39b60e645abb6fabb9769554829