Updated openstack/openstack Project: openstack-infra/puppet-log_processor  1f0f91fdbbaf5de145d8dfa471a943e3b8f248d7
Reduce log client logging by default

We had been running at debug level which is incredibly verbose. Remove
the -d flag. This will cause the logs which are logged to go to
stdout/err which should mean that upstart (or whatever init system) will
deal with them for us.

We should properly clean this up so that debug logging is useful again
in the long term.

Change-Id: I613c135ea56507d083df8c66e8846c6fbfa8b2ed