Updated openstack/openstack Project: openstack/nova  25be6a1e13f34774891fcfbdafe90d346b154852
libvirt: pick future min libvirt/qemu versions

The libvirt driver currently requires libvirt 1.2.1
and QEMU 1.5.3. In the Newton cycle we did not
announce any version bump for Ocata, so that will
not change.

This patch announces a version bump that will be
done in the Pike release, with predicted min versions
setting libvirt to 1.2.9 and QEMU to 2.1.0. These
are the versions present in Debian Jessie.

Out of the major distros currently supported, this
would eliminate support for:

 - Ubuntu Trusty. Workaround: enable the "Cloud Archive"
   the addon repository
 - SLES 12. Workaround: upgrade to 12SP1
 - RHEL 7.1. Workaround: upgrade to 7.2 or newer

Change-Id: I4a644aabf193d79c81e8fdae46e5def3526f78d4