Updated openstack/openstack Project: openstack/nova  f34b1c911a4b59724d37b6e0b748ce9d405d62a5
Fix database poison warnings, part 17

The following warning appears in the unit test logs a number of times.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

This patch fixes all the warnings from:

    nova.tests.unit.api.openstack.compute.test_serversV21.py

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I88264e54fd135e56612ae3066e1c160f6643f9c4
Related-Bug: #1568414