Updated openstack/openstack Project: openstack/nova  12ac2e76c2a021108837886765272388bb1acb6a
Fix database poison warnings, part 15

The following warning appears in the unit test logs a number of times.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

This patch fixes all the warnings from:

    unit.compute.test_compute_mgr.py

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I4529b4bdc13b4079b1274034b44a1f44922c3bf2
Related-Bug: #1568414