Updated openstack/openstack Project: openstack/cinder  25108d9743d202ff5b15fe7afd75d7f3b208332f
OVO: OPTIONAL_FIELDS should be moved inside the OVO

In working patch 251989, Michal Dulko points out that
all the OVOs should have OPTIONAL_FIELDS inside object.

This patch will fix the issue in OVO, group_type.

Now that the OPTIONAL_FIELDS is part of the class definition,
_from_db_object should no longer be a static method,
but should be a class method.

Change-Id: I619312761e92bc3e12d992657b7ad88331cdb060