Updated openstack/openstack Project: openstack/python-openstackclient  d0ecb8981d233d7ab7dcf52ee8bb0ca4e2e79f17
Add option "--name" to command "openstack object create"

Option "--name" can be used to set as the object name of
the file to be uploaded in the container. Similar to option
"--object-name" in command "swift upload". Added unit test case
to ensure an exception is raised when using option "--name" for
uploading multiple objects.

Change-Id: Ied7827841f6ca1cf9d4b48e304cbe5d62eda38ab
Closes-Bug: #1607972