Updated openstack/openstack Project: openstack/nova  b722632bb7c16ac7b60ab36efd4f5bc1ad3af057
Stop overwriting thread local context in ClientRouter

In commit 4df0869, a ClientRouter was added to cache client connections
to cell message queues. The periodic task for removing stale clients
is called with an empty RequestContext, but this overwrites the copy
in thread local storage unless overwrite=False is specified.

This adds overwrite=False to the empty RequestContext to prevent it
from overwriting the thread local context. All of the other periodic
tasks use get_admin_context() which also creates a RequestContext with
overwrite=False.

Closes-Bug: #1627838

Change-Id: I48024952865fe017e0cb786567b5b445b05e7659