Updated openstack/openstack Project: openstack/neutron-lbaas  ebf96b51a39f08ffba942cd3c5410265f5c57937
Use a cached get-pip.py if it is availble

Diskimage-builder has an element called pip-and-virtualenv that
specifies https://bootstrap.pypa.io/get-pip.py as the url to
dowload the get-pip.py file.  This patch checks if a version
is cached locally by infra on the gate hosts and will use that
location instead of downloading it.

Change-Id: I6f071984c1b26498223fc9a6f3528280ab064244