Updated openstack/openstack Project: openstack/cinder  3f602d3ea0d26c5b5e7ae1303b8aec1c00072b90
Remove dead code - replication RPC APIs

The REST APIs for promote_replica and reenable_replication were removed
in If862bcd18515098639f94a8294a8e44e1358c52a with implementation of
replication V2.1. Some code was left behind and now can be considered
dead. This commit removes it.

This should be safe from upgradeability point of view as we've deleted
the APIs back in Mitaka and we're supporting backward compatibility with
only one previous release.

Changes in the drivers are left for the driver maintainers.

Change-Id: Ia7dfcaa7601176ecb3a4b19bb59f6d9329e62635
Closes-Bug: 1620001