Updated openstack/openstack Project: openstack-infra/project-config  e194ac9c960ee92ad41d24bde422147731a5d4fa
Make docs non-voting for quark, remove template

Currently, the quark project does not have documentation, so
the repo lacks the structure to support it. This job has now
become a blocker for PR's to the project. Since it is a part
of the python-jobs template, remove that template. Re-add
needed gates, such as pep8 and python27.

Change-Id: Iedb700ebe8b1c379883652fd461fdbf2086f7bf4