Updated openstack/openstack Project: openstack/tempest  4a9e99109460a38f79715e04a241416c5a4930c5
Use ConfigParser instead of SafeConfigParser

SafeConfigParser supports interpolation on top of ConfigParser in Python 2,
and SafeConfigParser is deprecated in Python 3.2 and log warning
like " DeprecationWarning: The SafeConfigParser class has been renamed to
ConfigParser in Python 3.2. This alias will be removed in future versions.
Use ConfigParser directly instead."
So we can use ConfigParser if we don't need interpolation.

Change-Id: If210bd1322df96c0c8aea1f257148dc46affb422
Closes-Bug: #1618666