Updated openstack/openstack Project: openstack/oslo.policy  a3fe9b2d8ba381f5754e7d01d8bf18ef19eebde8
Update docs on policy sample generator

The docs for the policy sample generator were inconsistent with how it
is typically used in practice, with just one namespace. This led to
questions about using one namespace when in reality that's what happens.
This patch brings the example more in line with reality.

Change-Id: Id3c21459eff09aaf6e31c6f3aa4f8d829f87c513