Updated openstack/openstack Project: openstack/nova  00b349ba0978b9e6444fffdb2843eea4257e60ab
libvirt: Use the recreated disk.config.rescue during a rescue

I931421ea introduced a callback to _create_configdrive from
_create_domain during a rescue that recreates the config disk of an
instance with an additional .rescue suffix. However due to a latent bug
neither the original or this recreated config disk are mapped to the
instance during the resulting rescue.

To correct this disk.config.rescue is now added to the disk_mapping for
the instance if a config disk is required during a rescue.

Closes-bug: #1597669
Change-Id: I46cebd0612f50ff7373f677c4a002d444a877972