Updated openstack/openstack Project: openstack/openstack-manuals  11bf10a5f5b505e1be39712b3a4e92edb024c89e
Ubuntu mariadb install no longer prompts for pass

Installing Newton RC on Ubuntu 16.04,

apt-get install mariadb-server python-pymysql

no longer results in a prompt to set the root password,
bringing it into line with other distributions.

The install guide contains a reference to the previous behaviour

"Choose a suitable password for the database root account."

which is no longer necessary, and as such this patch removes it.

Change-Id: I46e8e02d7de00b08dff4648fe51a1da10bd053b9