Updated openstack/openstack Project: openstack-infra/storyboard-webclient  a79c818582c1e6b7475da5ba4fbf2082beacdfe3
Make validation match required project-title length

When creating a new project, StoryBoard asks the user
for a project title. Once that title is over 3 characters in length,
a big green tick is placed next to it. However, the title needs to
be a minimum of 5 characters in order for the project to actually
be saved. This commit makes the big green tick only appear once
a title is 5 characters or more in length.

Change-Id: Icbbe46ba6ef5abd9d818532761bf09cae675b0b3