Updated openstack/openstack Project: openstack-infra/storyboard-webclient  29aa6271d49741b52a52ec9b5fcb898e526d925a
Hide arrows to expand task details if there are no details

This isn't elegant, but it works. Previously, pages would
display the 'more details' arrows for tasks, even if there
were no more details. Users would click them, and nothing would
happen. This hides the arrows in those cases. In practice, the only
time there are no further details for a task is when a user is
logged out and the task has no notes. Should further fields display
in the task details for non-logged-in users, this code will need to
change to accommodate them.

Change-Id: I31e5b6cdc71dd4f362446557aa97d47381d8812a