Updated openstack/openstack Project: openstack/keystone  9534e6ffd94ffb87806cc9d6a4d6eafc139e2310
Make test_v3_auth exercise the whole API

Throughout much of the v3 authentication tests, you'd see calls
directly into the token_provider_api. While this was convenient
because the backends were loaded automatically, it makes changing
the implementation of the backend harder because we're relying on
it in the tests by coding to it directly, instead of just testing
the behavior through the API.

This commit makes it so that all calls to validate a token actually
use the token API instead of passing tokens directly to the
token_provider_api, exercising a request from the controller to the
backend... no cheat days here!

Change-Id: I5e792831dc808270f91f2c343e93e44a653b8676