Updated openstack/openstack Project: openstack/python-openstackclient  15069ef50e609dca769d6a50f6af92b2c8211238
Support "--no" option in aggregate set

Supporting "--no-property" option will apply user a convenience
way to clean all properties of aggregate in a short command,
and this kind of behavior is the recommended way to devref.
The patch add "--no-property" option in "aggregate set" command,
and update related test cases and devref document.

Change-Id: I7614a23c0db05144562330dc600dbab7d003d5d8
Implements: blueprint support-no-property-in-aggregate