Updated openstack/openstack Project: openstack/tempest  20e3d91ec32f0e9f8c92a5b8b0350865b21e07df
Fix tempest init inconsistency when it fails

This commit fixes an inconsistent behavior when tempest init fails.
For example::

  $ tempest init /foobar
  ERROR tempest [-] [Errno 13] Permission denied: '/foobar'
  $ tempest init /foobar
  A workspace already exists with name: foobar.
  $ tempest workspace list
  +------+------+
  | Name | Path |
  +------+------+
  +------+------+

The error message indicates that the workspace already exists but we
cannot see it with the tempest workspace list command. I think users
might be confused with this behavior.

And this commit also fixes --show-global-config-dir,-s option behavior
as a side effect. The original behavior was registering a current
directory as a workspace when the option was specified. However, I
think users don't expect the behavior.

Change-Id: I04ed102a4ad3c3aa678128f0004eb0ba09a05ea7